Extract the run_create method from Domain
The same run create method is to be used to make the `Machine`. The only
difference is the `deployment_parameters` and the `template_path`.

The resource_group is also required, however `machines` will use the
same group as their Domain. As such the code to create the group still
needs to live in the `Domain` model
