+++ lib/cloudware/providers/azure/domain.rb
@@ -15,18 +15,15 @@ module Cloudware
         private
 
         include Helpers::Client
+        include Helpers::Deploy
 
-        def run_create
-          client.resource.deployments.create_or_update(
-            resource_group.name, name, deployment_model
-          )
-        rescue MsRestAzure::AzureOperationError => e
-          # Azure returns a `JSON` string which contains an embedded `JSON`
-          # string. This embedded `JSON` contains the error message
-          message = JSON.parse(
-            JSON.parse(e.message)['response']['body']
-          )['error']['message']
-          raise InvalidAzureRequest, message
+        def deployment_parameters
+          {
+            cloudwareDomain: name,
+            cloudwareId: id,
+            networkCIDR: networkcidr,
+            priSubnetCIDR: prisubnetcidr
+          }
         end
 
         def run_destroy
@@ -46,39 +43,12 @@ module Cloudware
         end
         memoize :resource_group
 
-        def deployment_model
-          client.resource.model_classes.deployment.new.tap do |deployment|
-            deployment.properties = deployment_properties
-          end
-        end
-        memoize :deployment_model
-
-        def deployment_properties
-          client.resource.model_classes.deployment_properties.new.tap do |p|
-            p.template = template_content
-            p.parameters = convert_params_to_azure_syntax(
-              cloudwareDomain: name,
-              cloudwareId: id,
-              networkCIDR: networkcidr,
-              priSubnetCIDR: prisubnetcidr
-            )
-            p.mode = Azure::Resources::Profiles::Latest::Mgmt::Models::DeploymentMode::Incremental
-          end
-        end 
-
-        # Azure requires the parameter hash to have syntax:
-        # { :your_key => { value: 'your_value' } }
-        def convert_params_to_azure_syntax(**params)
-          params.map { |k, v| [k, { value: v }] }.to_h
-        end
-
-        def template_content
-          JSON.parse(File.read(File.join(
+        def template_path
+          File.join(
             Cloudware.config.base_dir,
             "providers/azure/templates/#{template}.json"
-          )))
+          )
         end
-        :memoize
       end
     end
   end
